refactor: 删除 pivot 中一些冗余的代码，将函数放置于合适的文件中
